title_search/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.21.0

